refactor(timetree): extract convergence and confidence domain logic from commands
- Domain computations (likelihood, ancestral state diffing, convergence metrics, rate susceptibility, CI extraction, quadrature combination) belong in the library layer for non-CLI consumers
- CLI orchestration (TimetreeOptimizer, TreetimeOptimizerTraceCsvWriter) and I/O (write_confidence_intervals) stay in commands/
