+++ packages/treetime/src/commands/timetree/convergence/mod.rs
@@ -1,6 +1,4 @@
-pub mod likelihood;
 pub mod metrics;
-pub mod sequence_changes;
 
 #[cfg(test)]
 mod __tests__;
