+++ packages/treetime/src/commands/timetree/refinement.rs
@@ -3,11 +3,11 @@ use crate::clock::clock_model::ClockModel;
 use crate::clock::clock_regression::{ClockParams, estimate_clock_model_with_reroot};
 use crate::clock::find_best_root::params::BranchPointOptimizationParams;
 use crate::commands::timetree::args::TreetimeTimetreeArgs;
-use crate::commands::timetree::convergence::sequence_changes::{capture_ancestral_states, count_sequence_changes};
 use crate::partition::timetree::GraphTimetree;
 use crate::partition::traits::PartitionTimetreeAll;
 use crate::payload::timetree::EdgeTimetree;
 use crate::payload::timetree::NodeTimetree;
+use crate::timetree::convergence::sequence_changes::{capture_ancestral_states, count_sequence_changes};
 use crate::timetree::inference::runner::run_timetree;
 use crate::timetree::optimization::polytomy::{prepare_tree_after_topology_change, resolve_polytomies};
 use crate::timetree::optimization::relaxed_clock::apply_relaxed_clock;
