+++ packages/treetime/src/mugration/mugration.rs
@@ -2,11 +2,11 @@ use crate::ancestral::marginal::update_marginal_mut;
 use crate::constants::MIN_BRANCH_LENGTH_FRACTION;
 use crate::gtr::gtr::{GTR, GTRParams};
 use crate::gtr::refinement::refine_gtr_iterative;
+use crate::make_error;
 use crate::mugration::input::MugrationInput;
 use crate::mugration::result::MugrationResult;
 use crate::partition::discrete_states::DiscreteStates;
 use crate::partition::marginal_discrete::PartitionMarginalDiscrete;
-use crate::make_error;
 use eyre::Report;
 use indexmap::IndexSet;
 use itertools::Itertools;
