+++ packages/treetime/src/timetree/mod.rs
@@ -1,3 +1,5 @@
+pub mod confidence;
+pub mod convergence;
 pub mod inference;
 pub mod optimization;
 pub mod utils;
